refactor: auth/domain 안에 있는 dto, mapper 패키지를 auth로 이동
