+++ src/main/java/com/upstage/devup/auth/controller/LogoutController.java
@@ -1,7 +1,7 @@
 package com.upstage.devup.auth.controller;
 
 import com.upstage.devup.auth.config.AuthenticatedUser;
-import com.upstage.devup.auth.domain.dto.LogoutResponseDto;
+import com.upstage.devup.auth.dto.LogoutResponseDto;
 import com.upstage.devup.global.exception.UnauthenticatedException;
 import com.upstage.devup.global.provider.CookieProvider;
 import jakarta.servlet.http.HttpServletResponse;
