+++ src/main/java/com/upstage/devup/auth/dto/LogoutResponseDto.java
@@ -1,4 +1,4 @@
-package com.upstage.devup.auth.domain.dto;
+package com.upstage.devup.auth.dto;
 
 public record LogoutResponseDto(String redirectUrl) {
 }
