+++ src/main/java/com/upstage/devup/auth/dto/SignInResult.java
@@ -1,4 +1,4 @@
-package com.upstage.devup.auth.domain.dto;
+package com.upstage.devup.auth.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
