+++ src/main/java/com/upstage/devup/auth/mapper/UserMapper.java
@@ -1,7 +1,7 @@
-package com.upstage.devup.auth.domain.mapper;
+package com.upstage.devup.auth.mapper;
 
-import com.upstage.devup.auth.domain.dto.SignUpRequestDto;
-import com.upstage.devup.auth.domain.dto.SignUpResponseDto;
+import com.upstage.devup.auth.dto.SignUpRequestDto;
+import com.upstage.devup.auth.dto.SignUpResponseDto;
 import com.upstage.devup.global.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
